requirements.txt
fastapi
uvicorn
jinja2
python-jose
reportlab
python-multipart
weasyprint
python-dotenv